css refactor, project directory restructured
