--- README.md
@@ -1,2 +1,0 @@
-# Snack-Overflow-Project1
-LAP 1 project 1 twitter clone - AKA snack overflow
