+++ frontend/src/containers/Main.js
@@ -1,4 +1,4 @@
-import Card from './Card';
+import Card from '../components/Card';
 import RenderCardList from './CardList';
 
 // class Post {
