+++ frontend/src/index.css
@@ -1,16 +1,25 @@
 /* Contents
 
-1) Variables
-    > Colors
+0) Normalise
+1) Colors 
 2) Typography
 3) Cards
-4) Media Queries
+4) Navbar
+5) Media Queries
 
+/* 
+    NOTE: This CSS stylesheet is intended to overwrite Bootstrap 4
+    cdn: https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css
 */
 
-/* 1) Variables */
+/* 0) Normalise */
+* {
+    padding: 0;
+    margin: 0;
+}
+
+/* 1) Colors */
 :root {
-    /* Colors */
     --offwhite: #f8f8f8;
     --verylightgrey: #dfdfdf;
     --nicegreen: #17a740;
@@ -32,20 +41,15 @@ p, a {
 }
 
 /* 3) Cards */
-
-/* NOTE: this card class is overwriting Bootstrap */
 .card {
     height: 300px;
     width: 100%;
     border: 0;
-
-    /* border-radius: 10px; */
-    /* border: solid rgb(215, 215, 215) 15px; */
+    border-radius: 0;
     background-color: var(--verylightgrey);
 }
 
 .card-control-panel {
-    /* background-color: red; */
     width: 80px;
     background-color: rgba(255,255,255,0.5);
 }
@@ -76,13 +80,9 @@ p, a {
 
 .card-control-panel p {
     font-size: 28px;
-    /* transform: rotate(-18deg);
-    margin-top: 15px;
-    margin-right: 12px; */
 }
 
 .card-content {
-    /* background-color: blue; */
     background: url("https://hips.hearstapps.com/hmg-prod.s3.amazonaws.com/images/190322-ham-sandwich-horizontal-1553721016.png?crop=0.668xw:1.00xh;0.168xw,0&resize=980:*");
     background-size: cover;
     background-position: center;
@@ -95,51 +95,22 @@ p, a {
     min-width: 375px;
 }
 
-/* Other */
+/* 4) Navbar */
 #navbar-icon {
     width: 30px;
     height: 30px;
 }
 
-/* .row {
- margin: 0;
- padding: 0;
- width: 100vw;
-}
-
-.container {
-    margin: 0;
-    padding: 0;
-}
-
-.App {
-    width: 100%;
-    height: 100%;
-}
-
-main {
-    padding: 0;
-    margin: 0;
-    width: 100%;
-} */
-
-/* 4) Media Queries */
-
-/*
-Breakpoints
-320px — 480px: Mobile devices.
-481px — 768px: iPads, Tablets.
-769px — 1024px: Small screens, laptops.
-1025px — 1200px: Desktops, large screens.
-1201px and more — Extra large screens, TV.
-*/
-
-@media screen and (min-width: 375px) {
+/* 5) Media Queries */
+/* small vertical margin between cards on larger screens */
+@media screen and (min-width: 574px) {
     .card {
-        margin: 20px 0;
+        margin: 15px 0;
     }
 }
-@media screen and (max-width: 375px) {
+
+/* remove margin at BS column breakpoint */
+@media screen and (max-width: 574px) {
     .card {
         margin: 0 0;
     }
